refactor: reorganize AI-related imports and remove deprecated files
