+++ lib/providers/ai-config.ts
