+++ lib/providers/ai-helpers.ts
